refactor: command 패키지 분리 및 안쓰는 코드 제거
