+++ 5-period/java-bridge-main/src/main/java/bridge/domain/bridge/Result.java
@@ -1,5 +1,7 @@
 package bridge.domain.bridge;
 
+import bridge.domain.command.UpDown;
+
 public class Result {
 
     private final UpDown upDown;
