--- 5-period/java-bridge-main/src/main/java/bridge/dto/MoveResultDto.java
@@ -1,9 +1,0 @@
-package bridge.dto;
-
-import bridge.domain.Direction;
-
-public record MoveResultDto(boolean isRightMove, char direction) {
-    public static MoveResultDto of(boolean isRightMove, Direction direction) {
-        return new MoveResultDto(isRightMove, direction.getDirection());
-    }
-}
